[ece4750-lab-server] Bug in finding available groups
Keys for groups dictionary are of type unicode/string; this due to how JSON stores its keys.
JSON only accept unicode/string for its keys and convert anything to string when storing.
Forgot to change the loop in finding available groups to look for a string of the group
number. This caused some people to end up "creating" the same group.

Also, returning a better and more descriptive message if the server cannot authenticate the
NetID.