[setup-ece4750] enforce source only one setup script, change prompt
This commit makes two changes.

1. We explicitly check to make sure no other course's setup script has
been sourced. This helps prevent conflicts where the tools from two
different courses conflict. Students will need to explicitly source the
appropriate setup script before working on a course, and log out and log
back in to work on a different course.

2. We used to not mess with a user's prompt, but students kept forgetting
to source the setup script. So now we change the prompt to make it obvious
that (1) they have sourced the setup script, and (2) what directory they
are currently working in.